Rejigged attacks to work with the dual-joystick system and not depend on frame rate

+++ app/src/main/java/org/o7planning/android2dgame/OrcAI.java
@@ -84,7 +84,7 @@ public class OrcAI implements CharacterAI {
                     character.hurtBox.x + character.hurtBox.width > other.hitBox.x &&
                     character.hurtBox.y < other.hitBox.y + other.hitBox.height &&
                     character.hurtBox.y + character.hurtBox.height > other.hitBox.y){
-                other.reduceHitPointsBy(character.attackDamage);
+                other.reduceHitPointsBy((double) character.attackDamage);
                 character.isAttacking = true;
             }
         }
