+++ app/src/main/java/org/o7planning/android2dgame/SlimeAI.java
@@ -53,7 +53,7 @@ public class SlimeAI implements CharacterAI {
                     character.hurtBox.x + character.hurtBox.width > other.hitBox.x &&
                     character.hurtBox.y < other.hitBox.y + other.hitBox.height &&
                     character.hurtBox.y + character.hurtBox.height > other.hitBox.y){
-                other.reduceHitPointsBy(character.attackDamage);
+                other.reduceHitPointsBy((double) character.attackDamage);
                 Log.d("Player HP remaining", ": " + other.hitPoints);
 
                 character.isAttacking = true;
